requirements.txt
psutil
patool~=2.2.0
py7zr~=0.21.0
maxminddb~=2.5.2
pyTelegramBotAPI~=4.16.1